Reduce warning. If there is more warning happening, I will fix later

+++ Super-Mario-DX10/Camera.cpp
@@ -17,7 +17,7 @@ void CCamera::SetCameraPosition(float marioX, int mapWidth) {
 
 	}
 	else if ((int)marioX + CGame::GetInstance()->GetScreenWidth() <= cameraFurthestPositionY) {
-		cameraPositionX = marioX - playerStartX;
+		cameraPositionX = (int) (marioX - playerStartX);
 
 	}
 	else {
@@ -42,7 +42,7 @@ void CCamera::SetCameraPosition(float marioX) {
 
 	}
 	else if ((int)marioX + CGame::GetInstance()->GetScreenWidth() <= cameraFurthestPositionY) {
-		cameraPositionX = marioX - playerStartX;
+		cameraPositionX = (int) (marioX - playerStartX);
 	}
 	else {
