+++ Super-Mario-DX10/HUD.cpp
@@ -331,7 +331,7 @@ void HUD::Update(DWORD dt, vector<LPGAMEOBJECT>* coObjects) {
 
 	//Update mario's power
 	if ((player->GetSpeedX() != 0 && CGame::GetInstance()->IsKeyDown(DIK_Z))|| power != 0) {
-		power = abs(player->GetSpeedX())* NUMBER_OF_POWER_ARROW /MARIO_SPEED_RUN_FLY_MAX;
+		power = (unsigned int) (abs(player->GetSpeedX())* NUMBER_OF_POWER_ARROW /MARIO_SPEED_RUN_FLY_MAX);
 	}
 	else {
 		power = 0;
@@ -370,7 +370,7 @@ void HUD::Update(DWORD dt, vector<LPGAMEOBJECT>* coObjects) {
 
 void HUD::Render() {
 	animation_set->at(0)->Render(this->x + CCamera::GetInstance()->GetCameraX(),this->y);
-	int i;
+	unsigned int i;
 	for (i = 0; i < POINT_DIGIT_NUMBER; i++) {
 		pointDigits[i]->Render();
 	}
